Make the authentication version number used accessible.
When creating an Auth struct to generate the verifier,
the version used is hardcoded to the latest version number.
But this version used was not visible to the client, and the client needed to
guess it in order to communicate it to the server.
We add the function `auth.GetVersion()` to make it accessible.